Time: 18 ms (5.72%) | Memory: 44.3 MB (26.69%) - LeetSync

+++ 326-power-of-three/power-of-three.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean isPowerOfThree(int n) {
-        for(long i =0;i<31;i++){
+        for(int i =0;i<31;i++){
             long ans =(long) Math.pow(3,i);
                 if(ans ==n){
                 return true;
